+++ .ccls-cache/@@home@runner@CPlusPlus-nix/@nix@store@iwd8ic5hhwdxn5dga0im55g5hjl270cd-glibc-2.33-47-dev@include@bits@stdlib-float.h
@@ -21,7 +21,7 @@
 #endif
 
 #ifdef __USE_EXTERN_INLINES
-__extern_inline double
+__extern_inline float
 __NTH (atof (const char *__nptr))
 {
   return strtod (__nptr, (char **) NULL);
