+++ .ccls-cache/@@home@runner@CPlusPlus/@nix@store@iwd8ic5hhwdxn5dga0im55g5hjl270cd-glibc-2.33-47-dev@include@time.h
@@ -75,7 +75,7 @@ extern clock_t clock (void) __THROW;
 extern time_t time (time_t *__timer) __THROW;
 
 /* Return the difference between TIME1 and TIME0.  */
-extern double difftime (time_t __time1, time_t __time0)
+extern float difftime (time_t __time1, time_t __time0)
      __THROW __attribute__ ((__const__));
 
 /* Return the `time_t' representation of TP and normalize TP.  */
